Moving load_tests to file.py
