+++ deps/wazuh_testing/wazuh_testing/tools/file.py
@@ -17,6 +17,7 @@ from os.path import exists
 
 import filetype
 import requests
+import yaml
 
 
 def read_json(file_path):
@@ -274,3 +275,16 @@ def set_file_owner_and_group(file_path, owner, group):
         gid = getgrnam(group).gr_gid
 
         os.chown(file_path, uid, gid)
+
+
+def load_tests(path):
+    """Loads a yaml file from a path.
+
+    Args:
+        path (string): Yaml full path.
+
+    Returns:
+       dict: Yaml structure.
+    """
+    with open(path) as f:
+        return yaml.safe_load(f)
