+++ tests/integration/test_authd/test_authd_key_hash.py
@@ -31,24 +31,13 @@ import yaml
 from wazuh_testing.tools import WAZUH_PATH
 from wazuh_testing.tools.configuration import load_wazuh_configurations
 from wazuh_testing.tools.services import control_service
+from wazuh_testing.tools.file import load_tests
 
 # Marks
 
 pytestmark = [pytest.mark.linux, pytest.mark.tier(level=0), pytest.mark.server]
 
 # Configurations
-
-
-def load_tests(path):
-    """Loads a yaml file from a path
-    Args:
-        path (str): path to the file.
-
-    Returns:
-        dict: dictionary containing the test info.
-    """
-    with open(path) as f:
-        return yaml.safe_load(f)
 
 
 test_data_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'data')
