+++ tests/integration/test_authd/test_authd_local.py
@@ -11,6 +11,7 @@ from wazuh_testing.tools import WAZUH_PATH
 from wazuh_testing.tools.configuration import load_wazuh_configurations
 # TODO Move to utils
 from wazuh_testing.tools.services import control_service
+from wazuh_testing.tools.file import load_tests
 
 # Marks
 
@@ -18,16 +19,6 @@ pytestmark = [pytest.mark.linux, pytest.mark.tier(level=0), pytest.mark.server]
 
 
 # Configurations
-
-def load_tests(path):
-    """ Loads a yaml file from a path
-    Returns
-    ----------
-    yaml structure
-    """
-    with open(path) as f:
-        return yaml.safe_load(f)
-
 
 test_data_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'data')
 message_tests = load_tests(os.path.join(test_data_path, 'local_enroll_messages.yaml'))
